main fix

+++ Assets/Code/Scripts/Player/Jump.cs
@@ -19,7 +19,7 @@ namespace Player.Controller
         private void Awake() => _character = GetComponent<Character2D>();
         private void Update()
         {
-            _character.animator.SetInteger("Gravity", (int)_character.body.linearVelocityY);
+            _character.animator.SetFloat("Gravity", _character.body.linearVelocityY);
             _isGrounded = Physics2D.BoxCast(transform.position, _size * Vector2.one, 0, Vector2.down, _distance, _groundMask);
             _coyoteJump = _character.body.linearVelocity.y <= 0 && _fallTime < _coyoteTime;
             _fallTime = _isGrounded ? 0 : _fallTime += Time.deltaTime;
